refactoring codebase
